+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [1.2.43] - 2021-03-15
+### Added
+- The `/api/playground_install` endpoint.
+- The `bold` and `italic` functions as Jinja2 filters.
+### Changed
+- As soon as the container starts, a dummy web server begins listening
+  on port 80 (or `PORT`).  This web server turns off before NGINX
+  starts.
+- Verbose `pip` logs no longer saved to `worker.log`.
+
 ## [1.2.42] - 2021-03-13
 ### Changed
 - The optional second positional parameter of `noun_plural()` and